factor out ssz helpers into new crate
