+++ coprocessor/Cargo.toml
@@ -8,11 +8,11 @@ name = "coprocessor"
 path = "src/main.rs"
 
 [dependencies]
-lazy_static="1.5"
+sha2.workspace=true
+ssz-merkleize.workspace=true
 serde = {version="1", features=["derive"]}
 hex = "0.4"
 tendermint-rpc = { version = "0.38.1", features = ["http-client"] }
-itertools = "0.14"
 ethereum-merkle-proofs = { package = "ethereum", git = "https://github.com/timewave-computer/zk-merkle-proofs", branch = "demo", features = [
     "borsh",
     "serde",
