+++ ssz-merkleize/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "ssz-merkleize"
+version = "0.1.0"
+edition = "2024"
+
+[dependencies]
+reqwest.workspace=true
+itertools.workspace=true
+sha2.workspace=true
+serde.workspace=true
+serde_json.workspace=true
+lazy_static="1.5.0"
